+++ app/gui/src/App.vue
@@ -1,13 +1,18 @@
 <script setup lang="ts">
 import LoadingScreenReact from '#/pages/authentication/LoadingScreen'
+import { EnsoPath } from '#/services/Backend'
 import RightPanel from '$/components/AppContainer/RightPanel.vue'
 import { useAppTitle } from '$/composables/appTitle'
+import { useAuth } from '$/providers/auth'
+import { provideContainerData } from '$/providers/container'
 import { provideOpenedProjects } from '$/providers/openedProjects'
 import { ContextsForReactProvider } from '$/providers/react/globalProvider'
+import { provideRightPanelData } from '$/providers/rightPanel'
+import { useText } from '$/providers/text'
 import ReactRoot from '$/ReactRoot'
 import { appOpenCloseCallback } from '$/utils/analytics'
 import '@/assets/base.css'
-import { interactionBindings } from '@/bindings'
+import { appBindings } from '@/bindings'
 import TooltipDisplayer from '@/components/TooltipDisplayer.vue'
 import { useEvent, useMounted } from '@/composables/events'
 import ProjectView from '@/ProjectView.vue'
@@ -27,10 +32,6 @@ import { Platform, platform } from 'enso-common/src/detect'
 import * as objects from 'enso-common/src/utilities/data/object'
 import { computed, onMounted, shallowRef } from 'vue'
 import { ComponentProps } from 'vue-component-type-helpers'
-import { useAuth } from './providers/auth'
-import { provideContainerData } from './providers/container'
-import { provideRightPanelData } from './providers/rightPanel'
-import { useText } from './providers/text'
 
 const { projectViewOnly } = defineProps<{
   // Used in Project View integration tests. Once both test projects will be merged, this should be
@@ -60,21 +61,19 @@ registerGlobalBlurHandler()
 
 const actionHandlers = registerHandlers(
   {
-    'interaction.cancel': { action: () => interaction.cancelAll() },
+    'app.cancel': { action: () => interaction.cancelAll() },
+    'app.close': { action: () => window.close() },
   },
   actions,
 )
 
-const interactionBindingsHandler = interactionBindings.handler(
-  objects.mapEntries(
-    interactionBindings.bindings,
-    (actionName) => actionHandlers[actionName].action,
-  ),
+const bindingsHandlers = appBindings.handler(
+  objects.mapEntries(appBindings.bindings, (actionName) => actionHandlers[actionName].action),
 )
 
 const { globalEventRegistry } = provideGlobalEventRegistry()
 
-useEvent(window, 'keydown', interactionBindingsHandler)
+useEvent(window, 'keydown', bindingsHandlers)
 useEvent(globalEventRegistry, 'pointerdown', (e) => interaction.handlePointerDown(e))
 
 const platformClass = (() => {
@@ -110,8 +109,8 @@ useMounted(appOpenCloseCallback)
 if (projectViewOnly) {
   const openedProjects = provideOpenedProjects()
   provideAsyncResources(openedProjects)
-  provideContainerData([])
-  provideRightPanelData(projectViewOnly.options.projectId, () => false, useText())
+  provideContainerData(EnsoPath(projectViewOnly.options.projectPath))
+  provideRightPanelData(EnsoPath(projectViewOnly.options.projectPath), () => false, useText())
   provideFullscreenRoot(fullscreenRoot)
 }
 </script>