+++ app/gui/src/authentication/service.ts
@@ -8,6 +8,7 @@ import * as appUtils from '$/appUtils'
 import * as cognitoModule from '$/authentication/cognito'
 import * as listen from '$/authentication/listen'
 import { useFeatureFlag } from '$/providers/featureFlags'
+import { parseEnsoDeeplink } from '@/util/url'
 import * as amplify from '@aws-amplify/auth'
 import * as common from 'enso-common'
 import type * as saveAccessTokenModule from 'enso-common/src/accessToken'
@@ -200,29 +201,79 @@ function loadAmplifyConfig(
  */
 function setDeepLinkHandler(logger: Logger, navigate: (url: string) => void) {
   window.authenticationApi.setDeepLinkHandler((urlString: string) => {
-    const url = new URL(urlString)
-    logger.log(`Parsed pathname: ${url.pathname}`)
-    // Remove the trailing slash in the pathname - it is present on Windows but not on macOS.
-    const pathname = url.pathname.replace(/\/$/, '')
-    switch (pathname) {
+    const result = parseEnsoDeeplink(urlString)
+    if (!result.ok) {
+      logger.log(result.error.message())
+      return
+    }
+    const deeplink = result.value
+    switch (deeplink.pathname) {
       // If the user is being redirected after clicking the registration confirmation link in their
       // email, then the URL will be for the confirmation page path.
-      case '//auth/confirmation': {
-        const verificationCode = url.searchParams.get('verification_code')
+      case 'auth/confirmation': {
+        const verificationCode = deeplink.searchParams.get('verification_code')
 
         let redirectUrl = ''
 
         // In case if the verifaction code is present, then we need to navigate to the confirmation
         // page, because the URL is a deep link for confirmation page and user is not yet confirmed.
         if (verificationCode != null) {
-          redirectUrl = `${appUtils.CONFIRM_REGISTRATION_PATH}${url.search}`
+          redirectUrl = `${appUtils.CONFIRM_REGISTRATION_PATH}${deeplink.search}`
         } else {
           // Otherwise, we need to navigate to the setup page, because user is already confirmed.
           // but the redirect link navigates to the confirmation page, for some reason.
-          redirectUrl = `${appUtils.DASHBOARD_PATH}${url.search}`
+          redirectUrl = `${appUtils.DASHBOARD_PATH}${deeplink.search}`
         }
         navigate(redirectUrl)
 
+        break
+      }
+      case 'auth': {
+        if (deeplink.search === '') {
+          // Signing out.
+          navigate(appUtils.LOGIN_PATH)
+        } else {
+          // Signing in.
+          void (async () => {
+            // Temporarily override the `history` object so that Amplify doesn't try to call
+            // `history.replaceState` (which doesn't work in the renderer process because of
+            // Electron's `webSecurity`). This is a hack, but it is the only way to get Amplify to
+            // work with a custom URL protocol in Electron.
+            // `history.replaceState` is only being saved here to be restored later.
+            // It will never be called without a bound `this`.
+            const replaceState = history.replaceState
+            history.replaceState = () => false
+            try {
+              // `_handleAuthResponse` is a private method without typings.
+              await amplify.Auth['_handleAuthResponse'](urlString)
+
+              navigate(appUtils.DASHBOARD_PATH)
+            } finally {
+              // Restore the original `history.replaceState` function.
+              history.replaceState = replaceState
+            }
+          })()
+        }
+        break
+      }
+      // If the user is being redirected after finishing the password reset flow, then the URL will
+      // be for the login page.
+      case 'auth/login': {
+        navigate(appUtils.LOGIN_PATH)
+        break
+      }
+      case 'auth/registration': {
+        navigate(`${appUtils.REGISTRATION_PATH}${deeplink.search}`)
+        break
+      }
+      case 'payments/success': {
+        navigate(`${appUtils.PAYMENTS_SUCCESS_PATH}${deeplink.search}`)
+        break
+      }
+      // If the user is being redirected from a password reset email, navigate to the password
+      // reset page, with the verification code and email prefilled.
+      case 'password-reset': {
+        navigate(`${appUtils.RESET_PASSWORD_PATH}${deeplink.search}`)
         break
       }
       // If the users is being redirected back from signing out from microsoft entra then we need to
@@ -239,56 +290,8 @@ function setDeepLinkHandler(logger: Logger, navigate: (url: string) => void) {
         })()
         break
       }
-      case '//auth': {
-        if (url.search === '') {
-          // Signing out.
-          navigate(appUtils.LOGIN_PATH)
-        } else {
-          // Signing in.
-          void (async () => {
-            // Temporarily override the `history` object so that Amplify doesn't try to call
-            // `history.replaceState` (which doesn't work in the renderer process because of
-            // Electron's `webSecurity`). This is a hack, but it is the only way to get Amplify to
-            // work with a custom URL protocol in Electron.
-            // `history.replaceState` is only being saved here to be restored later.
-            // It will never be called without a bound `this`.
-            const replaceState = history.replaceState
-            history.replaceState = () => false
-            try {
-              // @ts-expect-error `_handleAuthResponse` is a private method without typings.
-              await amplify.Auth._handleAuthResponse(url.toString())
-
-              navigate(appUtils.DASHBOARD_PATH)
-            } finally {
-              // Restore the original `history.replaceState` function.
-              history.replaceState = replaceState
-            }
-          })()
-        }
-        break
-      }
-      // If the user is being redirected after finishing the password reset flow, then the URL will
-      // be for the login page.
-      case '//auth/login': {
-        navigate(appUtils.LOGIN_PATH)
-        break
-      }
-      case '//auth/registration': {
-        navigate(`${appUtils.REGISTRATION_PATH}${url.search}`)
-        break
-      }
-      case '//payments/success': {
-        navigate(`${appUtils.PAYMENTS_SUCCESS_PATH}${url.search}`)
-        break
-      }
-      // If the user is being redirected from a password reset email, navigate to the password
-      // reset page, with the verification code and email prefilled.
-      case appUtils.RESET_PASSWORD_PATH: {
-        navigate(`${appUtils.RESET_PASSWORD_PATH}${url.search}`)
-        break
-      }
       default: {
-        navigate(pathname.slice(1))
+        navigate(`/${deeplink.pathname}`)
         break
       }
     }