+++ app/gui/src/dashboard/pages/dashboard/components/TheModal.tsx
@@ -1,12 +1,12 @@
 /** @file A component that renders the modal instance from the modal React Context. */
 import { Pressable } from '#/components/aria'
 import { Dialog } from '#/components/Dialog'
-import * as modalProvider from '#/providers/ModalProvider'
+import { useModal } from '#/providers/ModalProvider'
 import { AnimatePresence, motion } from 'framer-motion'
 
 /** Renders the modal instance from the modal React Context (if any). */
 export default function TheModal() {
-  const { modal, key } = modalProvider.useModal()
+  const { modal, key } = useModal()
 
   return (
     <AnimatePresence>