+++ app/gui/src/dashboard/services/RemoteBackend.ts
@@ -185,10 +185,14 @@ export default class RemoteBackend extends Backend {
 
   /**
    * Delete a user.
-   * FIXME: Not implemented on backend yet.
    */
-  override async removeUser(): Promise<void> {
-    return await this.throw(null, 'removeUserBackendError')
+  override async removeUser(userId: backend.UserId): Promise<void> {
+    const response = await this.delete(remoteBackendPaths.removeUserPath(userId))
+    if (!response.ok) {
+      return await this.throw(response, 'removeUserBackendError')
+    } else {
+      return
+    }
   }
 
   /** Invite a new user to the organization by email. */
@@ -803,15 +807,11 @@ export default class RemoteBackend extends Backend {
    * @throws An {@link DirectoryDoesNotExistError} if the asset is a directory and does not exist.
    * @returns The asset details. Returns `null` if the asset is a root directory.
    */
-  override async getAssetDetails<
-    Id extends backend.RealAssetId,
-    Type extends backend.RealAssetTypeId<Id>,
-    ReturnType extends Id extends backend.DirectoryId ?
-      backend.Asset<backend.AssetType.directory> | null
-    : backend.Asset<Type>,
-  >(assetId: Id): Promise<ReturnType> {
+  override async getAssetDetails<Id extends backend.RealAssetId>(
+    assetId: Id,
+  ): Promise<backend.AssetDetailsResponse<Id>> {
     const path = remoteBackendPaths.getAssetDetailsPath(assetId)
-    const response = await this.get<backend.Asset<Type> | null>(path)
+    const response = await this.get<backend.AssetDetailsResponse<Id>>(path)
 
     if (!response.ok) {
       if (response.status === STATUS_NOT_FOUND) {
@@ -825,8 +825,7 @@ export default class RemoteBackend extends Backend {
       return await this.throw(response, 'getAssetDetailsBackendError')
     }
 
-    // eslint-disable-next-line no-restricted-syntax
-    return (await response.json()) as ReturnType
+    return await response.json()
   }
   /**
    * Return Language Server logs for a project session.
@@ -1227,6 +1226,36 @@ export default class RemoteBackend extends Backend {
     }
   }
 
+  /**
+   * Fetches a configuration for a payment pricing page.
+   * @throws An error if a non-successful status code (not 200-299) was received.
+   */
+  async getPaymentsConfig(): Promise<backend.PaymentsConfig> {
+    const response = await this.get<backend.PaymentsConfig>(remoteBackendPaths.PAYMENTS_CONFIG_PATH)
+
+    if (!response.ok) {
+      return await this.throw(response, 'getPaymentsConfigBackendError')
+    } else {
+      return await response.json()
+    }
+  }
+
+  /**
+   * Cancel given subscription.
+   * @throws An error if a non-successful status code (not 200-299) was received.
+   */
+  override async cancelSubscription(subscriptionId: backend.SubscriptionId): Promise<void> {
+    const response = await this.delete(
+      remoteBackendPaths.cancelSubscriptionPath(subscriptionId),
+      {},
+    )
+    if (!response.ok) {
+      return await this.throw(response, 'cancelSubscriptionBackendError')
+    } else {
+      return
+    }
+  }
+
   /** List events in the organization's audit log. */
   override async getLogEvents(
     params: backend.GetLogEventsRequestParams,
@@ -1349,7 +1378,7 @@ export default class RemoteBackend extends Backend {
   async downloadProject(id: backend.ProjectId) {
     /** The type of the response body of this endpoint. */
     interface ResponseBody {
-      readonly targetDirectory: string
+      readonly projectRootDirectory: string
       readonly parentDirectory: string
     }
     const details = await this.getProjectDetails(id, true)
@@ -1365,7 +1394,7 @@ export default class RemoteBackend extends Backend {
     })
 
     const response = await this.client.get<ResponseBody>(
-      `./api/cloud/download-project?${queryString}`,
+      `/api/cloud/download-project?${queryString}`,
     )
     if (!response.ok) {
       return await this.throw(response, 'resolveProjectAssetPathBackendError')
@@ -1374,7 +1403,7 @@ export default class RemoteBackend extends Backend {
     const responseBody = await response.json()
 
     return {
-      targetId: backend.DirectoryId(`directory-${responseBody.targetDirectory}`),
+      projectRootId: backend.DirectoryId(`directory-${responseBody.projectRootDirectory}`),
       parentId: backend.DirectoryId(`directory-${responseBody.parentDirectory}`),
     }
   }
@@ -1385,7 +1414,7 @@ export default class RemoteBackend extends Backend {
       directory: extractIdFromDirectoryId(directoryId),
     })
 
-    const response = await this.client.get(`./api/cloud/get-project-archive?${queryString}`)
+    const response = await this.client.get(`/api/cloud/get-project-archive?${queryString}`)
     if (!response.ok) {
       return await this.throw(response, 'resolveProjectAssetPathBackendError')
     }
@@ -1410,8 +1439,11 @@ export default class RemoteBackend extends Backend {
   }
 
   /** Resolve asset metadata from an enso path. */
-  async resolveEnsoPath(path: backend.EnsoPath): Promise<backend.PathResolveResponse> {
-    const response = await this.get<backend.Asset>(remoteBackendPaths.RESOLVE_ENSO_PATH, { path })
+  override async resolveEnsoPath(path: backend.EnsoPath): Promise<backend.PathResolveResponse> {
+    const response = await this.get<backend.Asset<backend.RealAssetType>>(
+      remoteBackendPaths.RESOLVE_ENSO_PATH,
+      { path },
+    )
 
     if (!response.ok) return this.throw(response, 'resolveEnsoPathBackendError')
     return await response.json()