+++ app/gui/src/project-view/components/GraphEditor/GraphNode/nodeVisualization.ts
@@ -27,6 +27,7 @@ interface NodeVisualizationOptions {
   isFocused: ToValue<boolean>
   typeinfo: ToValue<Opt<TypeInfo>>
   dataSource: ToValue<Opt<VisualizationDataSource | RawDataSource>>
+  hidden: ToValue<boolean>
   emit: Emit
 }
 
@@ -39,6 +40,7 @@ export function useNodeVisualization({
   isFocused,
   typeinfo,
   dataSource,
+  hidden,
   emit,
 }: NodeVisualizationOptions) {
   const keyboard = injectKeyboard()
@@ -63,20 +65,17 @@ export function useNodeVisualization({
   const isVisualizationVisible = computed(
     () => isVisualizationEnabled.value || isVisualizationPreviewed.value,
   )
-  watch(isVisualizationVisible, (val) => {
-    // When visualization is being hidden, we don’t receive `pointerleave` event for some reason.
-    // So we need to set `visualizationHovered` to `false` manually.
-    if (!val) {
-      visualizationHovered.value = false
-    }
-  })
 
   const visRect = shallowRef<Rect>()
-  watch(visRect, (rect) => emit('update:visualizationRect', rect))
+  const visibleVisRect = computed(
+    (): Opt<Rect> => (isVisualizationVisible.value && !toValue(hidden) ? visRect.value : null),
+  )
+  watch(visibleVisRect, (rect) => emit('update:visualizationRect', rect ?? undefined))
 
   const visualization = computed((): ComponentProps<typeof GraphVisualization> => {
     const { size: nodeSize, pos: nodePosition } = toValue(nodeRect)
     return {
+      show: isVisualizationVisible.value,
       width: visualizationWidth.value,
       nodeSize,
       scale: toValue(scale),
@@ -103,7 +102,7 @@ export function useNodeVisualization({
     visualizationWidth,
     isVisualizationEnabled,
     isVisualizationPreviewed,
-    visRect: computed((): Opt<Rect> => (isVisualizationVisible.value ? visRect.value : null)),
-    visualization: computed(() => (isVisualizationVisible.value ? visualization.value : null)),
+    visRect: visibleVisRect,
+    visualization,
   }
 }