+++ CHANGELOG.md
@@ -29,6 +29,14 @@
 - ["Grouped Components" are renamed to "User Defined Components"][13389]
 - [Allow displaying cloud images using enso:// url in documentation][13419]
 - [File Browser Widget is used when editing file paths to datalinks][13439]
+- [When creating a node using the button on the port, the button can be dragged
+  to set its location][13598]
+- [Graph is not moved when showing/resizing side panels.][13557]
+- [Add "Invite" button to the top bar when using a team or higher plan][13522]
+- ["Welcome Project" is automatically opened for new users][13479]
+- [Project and Setting tab may be now closed with shortcut][13498][13604]. On
+  Windows/Linux <kbd>Ctrl</kbd>+<kbd>W</kbd> or <kbd>Ctrl</kbd> + <kbd>F4</kbd>;
+  on macOS: <kbd>⌘</kbd> + <kbd>W</kbd>.
 
 [12774]: https://github.com/enso-org/enso/pull/12774
 [12778]: https://github.com/enso-org/enso/pull/12778
@@ -52,6 +60,12 @@
 [13389]: https://github.com/enso-org/enso/pull/13389
 [13419]: https://github.com/enso-org/enso/pull/13419
 [13439]: https://github.com/enso-org/enso/pull/13439
+[13598]: https://github.com/enso-org/enso/pull/13598
+[13557]: https://github.com/enso-org/enso/pull/13557
+[13522]: https://github.com/enso-org/enso/pull/13522
+[13479]: https://github.com/enso-org/enso/pull/13479
+[13498]: https://github.com/enso-org/enso/pull/13498
+[13604]: https://github.com/enso-org/enso/pull/13604
 
 #### Enso Standard Library
 
@@ -72,6 +86,9 @@
 - [`Meta.Unresolved_Symbol` renamed to `Meta.Unresolved`][13443]
 - [Remane Google_Api library to Google][13436]
 - [Data.read_many now returns the read path as a char field][13475]
+- [Add prototype `find_group_number` function to Table][13487]
+- [Add support for custom and empty line endings to the fixed-width file format]
+  [13554]
 
 [12726]: https://github.com/enso-org/enso/pull/12726
 [12950]: https://github.com/enso-org/enso/pull/12950
@@ -85,18 +102,22 @@
 [13415]: https://github.com/enso-org/enso/pull/13415
 [13415]: https://github.com/enso-org/enso/pull/13443
 [13436]: https://github.com/enso-org/enso/pull/13436
+[13443]: https://github.com/enso-org/enso/pull/13443
 [13475]: https://github.com/enso-org/enso/pull/13475
+[13487]: https://github.com/enso-org/enso/pull/13487
+[13554]: https://github.com/enso-org/enso/pull/13554
 
 #### Enso Language & Runtime
 
 - [Allow optional path for `--jvm` option of `project-manager`][13225]
 - [Prevent `Meta` access to private constructors and private fields][12905]
-- [Encapsulating Private_Access constructor][#12976]
+- [Encapsulating Private_Access constructor][12976]
 - [Upgrading Truffle][12500] (including its
   [JavaScript](https://www.graalvm.org/javascript/) and
   [Python](https://www.graalvm.org/python/)) to version `24.2.0`.
 - [Upgrade GraalVM from JDK 21 to JDK 24][12855]
 - [Use JAVA_TOOL_OPTIONS env variable to alter JVM arguments][13256]
+- [Check private access at resolution, not on invocation][13392]
 
 [12500]: https://github.com/enso-org/enso/pull/12500
 [12976]: https://github.com/enso-org/enso/pull/12976
@@ -104,6 +125,7 @@
 [12905]: https://github.com/enso-org/enso/pull/12905
 [13225]: https://github.com/enso-org/enso/pull/13225
 [13256]: https://github.com/enso-org/enso/pull/13256
+[13392]: https://github.com/enso-org/enso/pull/13392
 
 # Enso 2025.1